Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun, Kerbol, A.K.A. the Sun,
Duna is the fourth planet in the Kerbol system and ranks as the fourth largest. Known as the Kerbal Space Program's counterpart to Mars, it features a red-brown surface and polar ice caps akin to those found on Kerbin. Duna hosts one prominent natural satellite, the tidally-locked moon called Ike.

Duna orbits Kerbol with an inclination nearly identical to that of Kerbin, making it one of the easiest planets to encounter. Achieving a round trip from Kerbin to orbit around Duna and back requires approximately 1700 m/s of delta-v from Low Kerbin Orbit (LKO), assuming proper aerobraking and a ballistic insertion burn from 100 km. Different orientations of Kerbin and Duna may necessitate higher delta-v budgets for the journey.

In-game description
KSP 1:

“	Also known as the red dot that you can see if you squint at it really hard, Duna has long been a wonder to Kerbalkind. The planet has been held in much awe, due to its striking red color and stark contrast to the color green.
— Kerbal Astronomical Society

”
KSP 2:

“	Duna, The Reddest Planet, has long been an object of study and inspiration to Kerbal scientists and artists alike! Its thin, low-pressure atmosphere allows for both aerobraking and relatively easy access to its surface. Kerbal researches and sci-fi writers often discuss the discoveries that could be waiting for them in the planet's polar ice caps or on Duna's tidally locked moon, Ike.

Topography

Duna's terrain ranges in altitude from 124 to 8264 meters. Its mountain ranges are a peril to landing craft, due to their steepness and the limited aerobraking that is possible in Duna's tenuous atmosphere at such altitudes.

Duna has several craters. They appear quite eroded, presumably by the wind of the (small) atmosphere.

Several maria (large regions of dark soil) are around several hundred meters altitude. Their low altitude allows for more aerobraking, and the flatness of the terrain makes them easy but boring landing locations. These may have been lakes when the planet may have had liquid on its surface in the past.

Duna's surface contains 5 anomalous features.

Biomes

Duna has 14 Biomes. Within the generic mix of Low-, Mid- and Highlands lie various Basins and Canyons with some scattered craters. The most dominant feature is the Midland Sea,[1] which spans half the Planet. The Poles also include separate Highlands and Craters. Duna also has gigantic ice caps.

Atmosphere

Duna has a thin, cool atmosphere with a mass of approximately 3.0×1015 kilograms, a datum level pressure of 6.755 kilopascals (0.067 atmosphere), and a depth of 50,000 meters. Compared to the atmosphere of Kerbin, Duna's atmosphere has 1/15th the mass, 1/15th the surface pressure, and 1/9th the surface density. The datum level pressure on Duna is equivalent to that at an altitude of 14,975 m on Kerbin. The pressure at the top of Duna's highest mountain peak is only 2.4 kPa.

The average molecular weight of Duna air is 42 g/mol, and its adiabatic index is 1.20. Although the composition of Duna's atmosphere is unknown, these values suggest that it may consist largely of carbon dioxide.

Like all other atmospheres in the game, Duna's atmosphere fades exponentially as altitude increases. The scale height varies with altitude, which is a change from pre-1.0 versions of the game. The pressure-altitude profile is globally constant and independent of temperature.

The surface of Duna is cold, with a globally averaged datum level temperature of approximately -23 °C. Air temperatures decrease with increasing altitude to an altitude of 30 km. Between the altitudes of 30-45 km lies an isothermal layer, having a temperature of -123 °C.

Air temperatures vary with latitude and time of day. At the equator, datum level temperatures vary between a nighttime low of -20 °C and a daytime high of -2 °C. At the poles, the temperature varies between -90 °C and -85 °C. Since Duna has no axial tilt, there are no seasonal temperature variations

Duna's atmosphere appears to have a reddish or brownish color, both from the surface and from space. The atmosphere has this color even over the poles, where the surface is white.

Orbital statistics
The synchronous orbit of Duna is at an altitude of 2 880.00 km above Duna. The speed of the satellite is 306.88 m/s and it has an orbital period resonating with 1 Duna day (18.2 hours or 65517.859375 seconds). However, this is the same altitude at which Ike orbits, so unless a ship's orbit is perfectly precise, it will eventually be captured and crash into Ike.

For a semi-synchronous orbit of ½ Duna day (9.1 hours or 32758.9295 seconds) an orbit of 1 695.87 km above Duna is needed with a velocity of 386.65 m/s.